refactor(bootstrap): move composition root out of socket adapter
Move GameContext from core.server.socket to a new io.taanielo.jmud.bootstrap
package via git mv, add a @NullMarked package-info, and update its Javadoc to
state it is the only place allowed to construct Json* repositories and
infrastructure schedulers. Dedupe repository construction in
GameContext.create(): JsonItemRepository, JsonAttackRepository,
JsonRaceRepository, and JsonClassRepository are now each built exactly once
via new createItemRepository/createAttackRepository/createRaceRepository/
createClassRepository helpers and threaded through all consumers, instead of
being constructed up to 4x each. Update all referencing imports (Main,
SocketClient, SocketServer, SshGameShellFactory, SshGameShell, SshServer) and
extend ArchitectureTest's transport_isolation exemption to ..bootstrap...

Pure mechanical move + dedup, no behavior change.

Closes #181
